requirements.txt
django==1.8.4
djangorestframework==3.2.2
